requirements.txt
fastapi
uvicorn
requests
python-dotenv
scikit-learn
pandas
numpy
geopygoogle-generativeai
